ex - consolidate common py code
